store measures and meas count
