+++ frontend/package.json
@@ -20,8 +20,6 @@
     "konva": "^8.3.14",
     "peaks.js": "3.0.0-beta.0",
     "pinia": "^2.0.13",
-    "uplot": "^1.6.26",
-    "uplot-vue": "^1.1.5",
     "vue": "^3.2.25",
     "vue-echarts": "^6.2.3",
     "vue-i18n": "^9.1.9",
