+++ frontend/src/components/feature_visualizer/javascript/variables.js
@@ -1,0 +1,40 @@
+import { computed, ref } from 'vue';
+
+const colors = [
+    '#1f77b4',
+    '#ff7f0e',
+    '#2ca02c',
+    '#d62728',
+    '#9467bd',
+    '#8c564b',
+    '#e377c2',
+    '#7f7f7f',
+    '#bcbd22',
+    '#17becf',
+];
+
+const currentMeasure = ref(-1);
+const cursorPositions = ref([]);
+const endMeasure = ref(0);
+const featureVisualizerOpened = ref(false);
+const isPlaying = ref(false);
+const labelSelectors = ref([]);
+const measureSelector = ref(null);
+const measuresVisible = ref(false);
+const peaksInstancesReady = ref([]);
+const playing = ref([]);
+const regionOverlay = ref([]);
+const selectedLabel = ref('');
+const startMeasure = ref(0);
+const timeSelections = ref([]);
+const trackLabels = ref([]);
+const trackNames = ref([]);
+const tracksVisible = ref([]);
+const volume = ref(1.0);
+const waveformsVisible = ref([]);
+
+const allPeaksReady = computed(() => {
+    return peaksInstancesReady.value.includes(false);
+});
+
+export { allPeaksReady, featureVisualizerOpened };
