+++ frontend/src/components/feature_visualizer/subcomponents/LineChart.vue
