+++ frontend/src/components/feature_visualizer/subcomponents/LineChartMeasure.vue
