+++ frontend/src/components/feature_visualizer/subcomponents/MeasureSelector.vue
